requirements.txt
tqdm
sympy
numpy
scipy
torch_scatter
omegaconf
pykdtree
ninja
matplotlib
pytorch_lightning==1.6.4

--extra-index-url http://eagle.huangjh.tech:8080/simple
--trusted-host eagle.huangjh.tech
python-pycg==0.2.0
